names.go
package json2go

import (
	"bytes"
	"strings"
)

// attrName converts json field name to pretty struct attribute name
func attrName(fieldName string) string {
	var b bytes.Buffer

	words := strings.Split(fieldName, "_")
	for _, w := range words {
		if u := strings.ToUpper(w); commonInitialisms[u] {
			b.WriteString(u)
			continue
		}

		b.WriteString(strings.Title(w))
	}

	return b.String()
}

// commonInitialisms is a set of common initialisms.
//
// source: https://github.com/golang/lint/blob/master/lint.go
var commonInitialisms = map[string]bool{
	"ACL":   true,
	"API":   true,
	"ASCII": true,
	"CPU":   true,
	"CSS":   true,
	"DNS":   true,
	"EOF":   true,
	"GUID":  true,
	"HTML":  true,
	"HTTP":  true,
	"HTTPS": true,
	"ID":    true,
	"IP":    true,
	"JSON":  true,
	"LHS":   true,
	"QPS":   true,
	"RAM":   true,
	"RHS":   true,
	"RPC":   true,
	"SLA":   true,
	"SMTP":  true,
	"SQL":   true,
	"SSH":   true,
	"TCP":   true,
	"TLS":   true,
	"TTL":   true,
	"UDP":   true,
	"UI":    true,
	"UID":   true,
	"UUID":  true,
	"URI":   true,
	"URL":   true,
	"UTF8":  true,
	"VM":    true,
	"XML":   true,
	"XMPP":  true,
	"XSRF":  true,
	"XSS":   true,
}
